refactor(hub): shared webhook signature helper in integrations
Extract the per-integration HMAC blocks into one shared helper
(verifyHMACSHA256 in signature.go) plus the constant-time shared-secret
compare used by Jira, and document the signature-validation audit for
all six webhook entry points (linear, github PR, github issues,
shortcut, jira, external). All integrations do validate signatures;
their differing no-secret defaults (GitHub PR strict, others open) are
preserved unchanged.
