+++ pkg/hub/integrations/github_webhook.go
@@ -3,9 +3,6 @@ package integrations
 import (
 	"bytes"
 	"context"
-	"crypto/hmac"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -166,10 +163,7 @@ func (s *Service) validateGitHubSignature(body []byte, sig string) bool {
 		if secret == "" {
 			continue
 		}
-		mac := hmac.New(sha256.New, []byte(secret))
-		mac.Write(body)
-		expected := hex.EncodeToString(mac.Sum(nil))
-		if hmac.Equal([]byte(sig), []byte(expected)) {
+		if verifyHMACSHA256(body, secret, sig) {
 			return true
 		}
 	}
